Added new Landing page for new users.
