+++ app/src/main/res/layout/fragment_first_slide.xml
@@ -42,6 +42,7 @@
             android:layout_gravity="center"
             android:paddingLeft="16dp"
             android:paddingRight="16dp"
+            android:tint="@android:color/white"
             android:src="@drawable/ic_quick_contacts_dialer_24dp"/>
     </LinearLayout>
     <TextView
